refactor(tests): split useWorkspaceWatch.test.ts into connection + messages
Split the 265-line monolithic test into two focused files:

- useWorkspaceWatch.connection.test.ts (4 tests, 114 lines)
  — WebSocket connection lifecycle, abort on unmount, toggle

- useWorkspaceWatch.messages.test.ts (4 tests, 4 describe blocks)
  — SSE data parsing, event filtering, multi-hook fanout

Each file is self-contained with its own imports, mocks and helpers.
All 8 original tests preserved with identical logic.

Addresses Medium review issue on PR #5409.
